+++ app/src/main/java/com/school_of_company/gwangsan/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.school_of_company.gwangsan_android.ui.theme
+package com.school_of_company.gwangsan.ui.theme
 
 import androidx.compose.ui.graphics.Color
