+++ app/src/main/java/com/school_of_company/gwangsan/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.school_of_company.gwangsan_android.ui.theme
+package com.school_of_company.gwangsan.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
